requirements.txt
yt-dlp
openai-whisper
torch
requests
python-dotenv
functions-framework
flask
colorama